Proposal: extract LVM space-making part
