+++ src/lib/y2storage/proposal/lvm_creator.rb
@@ -20,6 +20,7 @@
 require "y2storage/planned"
 require "y2storage/disk_size"
 require "y2storage/proposal/creator_result"
+require "y2storage/proposal/lvm_space_maker"
 
 module Y2Storage
   module Proposal
@@ -129,61 +130,13 @@ module Y2Storage
 
       # Makes space for planned logical volumes
       #
-      # When making free space, three different policies can be followed:
-      #
-      # * :needed: remove logical volumes until there's enough space for
-      #            planned ones.
-      # * :remove: remove all logical volumes.
-      # * :keep:   keep all logical volumes.
-      #
       # This method modifies the volume group received as first argument.
       #
       # @param volume_group [LvmVg] volume group to clean-up
       # @param planned_vg   [Planned::LvmVg] planned logical volume
       def make_space(volume_group, planned_vg)
-        return if planned_vg.make_space_policy == :keep
-
-        case planned_vg.make_space_policy
-        when :needed
-          make_space_until_fit(volume_group, planned_vg.lvs)
-        when :remove
-          lvs_to_keep = planned_vg.all_lvs.select(&:reuse?).map(&:reuse_name)
-          remove_logical_volumes(volume_group, lvs_to_keep)
-        end
-      end
-
-      # Makes sure the given volume group has enough free extends to allocate
-      # all the planned volumes, by deleting the existing logical volumes.
-      #
-      # This method modifies the volume group received as first argument.
-      #
-      # FIXME: the current implementation does not guarantee than the freed
-      # space is the minimum valid one.
-      #
-      # @param volume_group [LvmVg] volume group to modify
-      def make_space_until_fit(volume_group, planned_lvs)
-        space_size = DiskSize.sum(planned_lvs.map(&:min_size))
-        missing = missing_vg_space(volume_group, space_size)
-        while missing > DiskSize.zero
-          lv_to_delete = delete_candidate(volume_group, missing)
-          if lv_to_delete.nil?
-            error_msg = "The volume group #{volume_group.vg_name} is not big enough"
-            raise NoDiskSpaceError, error_msg
-          end
-          volume_group.delete_lvm_lv(lv_to_delete)
-          missing = missing_vg_space(volume_group, space_size)
-        end
-      end
-
-      # Remove all logical volumes from a volume group
-      #
-      # This method modifies the volume group received as a first argument.
-      #
-      # @param volume_group [LvmVg]         volume group to remove logical volumes from
-      # @param lvs_to_keep  [Array<String>] name of logical volumes to keep
-      def remove_logical_volumes(volume_group, lvs_to_keep)
-        lvs_to_remove = volume_group.all_lvm_lvs.reject { |v| lvs_to_keep.include?(v.name) }
-        lvs_to_remove.each { |v| volume_group.delete_lvm_lv(v) }
+        space_maker = LvmSpaceMaker.new(volume_group, planned_vg)
+        space_maker.provide_space
       end
 
       # Creates a logical volume for each planned volume.
@@ -230,32 +183,6 @@ module Y2Storage
         planned_lv.format!(new_lv)
         add_stripes_config(new_lv, planned_lv)
         new_lv
-      end
-
-      # Best logical volume to delete next while trying to make space for the
-      # planned volumes. It returns the smallest logical volume that would
-      # fulfill the goal. If no LV is big enough, it returns the biggest one.
-      def delete_candidate(volume_group, target_space)
-        lvs = volume_group.lvm_lvs
-        big_lvs = lvs.select { |lv| lv.size >= target_space }
-        if big_lvs.empty?
-          lvs.max_by(&:size)
-        else
-          big_lvs.min_by(&:size)
-        end
-      end
-
-      # Missing space in the volume group to fullfil a target
-      #
-      # @param volume_group [LvmVg]    Volume group
-      # @param target_space [DiskSize] Required space
-      def missing_vg_space(volume_group, target_space)
-        available = volume_group.available_space
-        if available > target_space
-          DiskSize.zero
-        else
-          target_space - available
-        end
       end
 
       # Returns the name that is available taking original_name as a base. If
